demo_applications/advance_rag/multimodal-rag/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.1
langchain==0.2.12
langchain-openai==0.1.7
chromadb==0.5.3
pydantic==2.8.2
torch==2.2.2
transformers==4.41.0
pillow==10.4.0
python-multipart==0.0.9
